Rename stepper component files to kebab-case
