+++ packages/js/src/first-time-configuration/tailwind-components/step-circle.js
