+++ packages/js/src/first-time-configuration/tailwind-components/step-header.js
@@ -1,4 +1,4 @@
-import { StepCircle } from "./StepCircle";
+import { StepCircle } from "./step-circle";
 import PropTypes from "prop-types";
 import { useEffect, useState } from "@wordpress/element";
 import { stepperTimings } from "../stepper-helper";
